+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project, organized by day.
 - Added color on icons.
 - Added custom confirm modal.
 - Removed open file functions.
+- Added new extensions and icons.
 
 ## 2026-09-24
 - Moved `publish_new_version.sh` in the `/scripts` folder.